refactor: implement framework registry
