+++ src/registries/backend.js
@@ -1,0 +1,5 @@
+import { setupExpressBackend } from '../stacks/backend/express.js';
+
+export const backendStacks = {
+  'express': { setup: setupExpressBackend }
+};
