+++ src/registries/ui.js
@@ -1,0 +1,7 @@
+import * as tailwind from '../ui/tailwind.js';
+import * as chakra from '../ui/chakra.js';
+
+export const uiFrameworks = {
+  'Tailwind': { install: tailwind.install },
+  'Chakra': { install: chakra.install }
+};
