+++ src/ui/tailwind.js
@@ -1,0 +1,4 @@
+import { setupUIFramework } from '../utils.js';
+export async function install(targetDir, stackType) {
+  await setupUIFramework(targetDir, 'Tailwind', stackType);
+}
